+++ _test/test/goldens/generated_build_script.dart
@@ -13,10 +13,9 @@ import 'package:build_modules/builders.dart' as _i5;
 import 'package:build_vm_compilers/builders.dart' as _i6;
 import 'package:build_web_compilers/builders.dart' as _i7;
 import 'package:build/build.dart' as _i8;
-import 'dart:async' as _i9;
-import 'dart:isolate' as _i10;
-import 'package:build_runner/build_runner.dart' as _i11;
-import 'dart:io' as _i12;
+import 'dart:isolate' as _i9;
+import 'package:build_runner/build_runner.dart' as _i10;
+import 'dart:io' as _i11;
 
 final _builders = <_i1.BuilderApplication>[
   _i1.apply(
@@ -139,8 +138,8 @@ final _builders = <_i1.BuilderApplication>[
   _i1.applyPostProcess(
       r'provides_builder:some_post_process_builder', _i4.somePostProcessBuilder)
 ];
-_i9.Future<void> main(List<String> args, [_i10.SendPort sendPort]) async {
-  var result = await _i11.run(args, _builders);
+void main(List<String> args, [_i9.SendPort sendPort]) async {
+  var result = await _i10.run(args, _builders);
   sendPort?.send(result);
-  _i12.exitCode = result;
+  _i11.exitCode = result;
 }